+++ src/OrchardCore.Modules/OrchardCore.Users/ViewModels/EnableAuthenticatorViewModel.cs
@@ -7,6 +7,7 @@ public class EnableAuthenticatorViewModel
     public string SharedKey { get; set; }
 
     public string AuthenticatorUri { get; set; }
+
     public string ReturnUrl { get; set; }
 
     [Required]
