+++ src/OrchardCore.Modules/OrchardCore.Users/Views/TwoFactorAuthentication/GenerateRecoveryCodes.cshtml
@@ -1,16 +1,9 @@
 <zone Name="Title"><h1>@RenderTitleSegments(T["Generate two-factor authentication (2FA) recovery codes"])</h1></zone>
 
 <div class="alert alert-warning" role="alert">
-    <p>
-    <p>
-        <strong>@T["Store these codes in a safe place."]</strong>
-    </p>
-    <p>
-            @T["If you lose your device and don't have the recovery codes you will lose access to your account."]"
-    </p>
-    <p>
-            @T["Generating new recovery codes does not change the keys used in authenticator apps. If you wish to change the key used in an authenticator app you should <a href=\"{0}\">reset your authenticator keys.", Url.Action("ResetAuthenticator")]</a>
-    </p>
+    <p><strong>@T["Store these codes in a safe place."]</strong></p>
+    <p>@T["If you lose your device and don't have the recovery codes you will lose access to your account."]"</p>
+    <p>@T["Generating new recovery codes does not change the keys used in authenticator apps. If you wish to change the key used in an authenticator app you should <a href=\"{0}\">reset your authenticator keys</a>.", Url.Action("ResetAuthenticator")]</p>
 </div>
 <div>
     <form method="post">
