+++ src/OrchardCore.Modules/OrchardCore.Users/Views/TwoFactorAuthentication/ShowRecoveryCodes.cshtml
@@ -3,12 +3,8 @@
 <zone Name="Title"><h1>@RenderTitleSegments(T["Recovery codes"])</h1></zone>
 
 <div class="alert alert-warning" role="alert">
-    <p>
-        <strong>@T["Store these codes in a safe place."]</strong>
-    </p>
-    <p>
-        @T["If you lose your device and don't have the recovery codes you will lose access to your account."]"
-    </p>
+    <p><strong>@T["Store these codes in a safe place."]</strong></p>
+    <p>@T["If you lose your device and don't have the recovery codes you will lose access to your account."]</p>
 </div>
 
 <div class="row">
