+++ src/OrchardCore.Modules/OrchardCore.Users/Views/UserMenu.cshtml
@@ -1,7 +1,7 @@
-@using System.Security.Claims
-@using OrchardCore.Settings
-@using OrchardCore.Users.Models
-@using OrchardCore.Entities
+@    using System.Security.Claims
+@    using OrchardCore.Settings
+@    using OrchardCore.Users.Models
+@    using OrchardCore.Entities
 
 @inject ISiteService SiteService
 @inject IAuthorizationService AuthorizationService
