+++ src/OrchardCore.Themes/TheTheme/Views/LoginMenu.cshtml
@@ -1,8 +1,8 @@
-@using Microsoft.AspNetCore.Identity
-@using OrchardCore.Entities
-@using OrchardCore.Settings
-@using OrchardCore.Users
-@using OrchardCore.Users.Models
+@    using Microsoft.AspNetCore.Identity
+@    using OrchardCore.Entities
+@    using OrchardCore.Settings
+@    using OrchardCore.Users
+@    using OrchardCore.Users.Models
 @inject ISiteService SiteService
 @inject SignInManager<IUser> SignInManager
